Add f64 function variants and DRY up guards (#55)
This commit adds several `f64` variants of calculation methods:

* `mean_f64`
* `stdvar_f64`
* `stddev_f64`

The bodies of the original functions are also replaced with calls to these
variants followed by `ceil()`. This is functionally equivalent to the previous
pattern, but avoids repetition. Explicit test cases are not added for the
variants because the `u64` test cases implicitly cover them.

The new variants should resolve #52.

The commit also refactors out a common guard, `entries` followed by a check
against the size, to a new function `try_entries`, and uses this in other
locations.

Last, requesting an invalid percentile value (that is, if `0.0 <= p <= 100.0`
doesn't hold) will now return a dedicated error type, `InvalidPercentileValue`,
rather than an `UnknownError`. Tests are added for this case.